// frontend/src/generated/types/menuDetailResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */

export interface MenuDetailResponse {
  sysId?: string;
  linkSysId?: string;
  menuName?: string;
  menuPrice?: number;
  menuDescription?: string;
  optionUseYn?: string;
  linkSysId2?: string;
  useYn?: string;
  fileUlid?: string;
  ordNo?: number;
}