// frontend/src/generated/types/menuOptionDetailRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */
import type { MenuOptionDetailItem } from './menuOptionDetailItem';

export interface MenuOptionDetailRequest {
  newItems?: MenuOptionDetailItem[];
  updateItems?: MenuOptionDetailItem[];
  delItems?: MenuOptionDetailItem[];
}